pyproject.toml
[tool.poetry]
name = "record"
version = "0.1.0"
description = ""
authors = ["Andrew Chen <andrew.chen.anyuan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.46.3"
Pillow = "^9.1.1"
google = "^3.0.0"
protobuf = "^4.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
